+++ Sources/Sentry/include/_SentryDispatchQueueWrapperInternal.h
@@ -12,6 +12,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property (strong, nonatomic) dispatch_queue_t queue;
 
+- (instancetype)initWithName:(const char *)name;
+
 - (instancetype)initWithName:(const char *)name
                   attributes:(nullable dispatch_queue_attr_t)attributes;