+++ Sources/Swift/Core/Helper/SentryFileContents.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 @objcMembers
-@_spi(Private) public class SentryFileContents: NSObject {
+@_spi(Private) public final class SentryFileContents: NSObject {
     
     public let path: String
     public let contents: Data