+++ Sources/Swift/Integrations/SessionReplay/Preview/SentryMaskingPreviewView.swift
@@ -3,7 +3,7 @@ import Foundation
 import UIKit
 
 @objcMembers
-@_spi(Private) public class SentryMaskingPreviewView: UIView {
+@_spi(Private) public final class SentryMaskingPreviewView: UIView {
     private class PreviewRenderer: SentryViewRenderer {
         func render(view: UIView) -> UIImage {
             return UIGraphicsImageRenderer(size: view.frame.size, format: .init(for: .init(displayScale: 1))).image { _ in